migrate custom pin to typescript
