--- legacy/components/FirstRun/styles/CustomPinScreen.style.js
@@ -1,11 +1,0 @@
-// @flow
-
-import {StyleSheet} from 'react-native'
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-  },
-})
-
-export default styles
