+++ src/Settings/CustomPin/CustomPinScreen.stories.tsx
@@ -1,11 +1,9 @@
-// @flow
-
 import {NavigationRouteContext} from '@react-navigation/native'
 import {action} from '@storybook/addon-actions'
 import {storiesOf} from '@storybook/react-native'
 import React from 'react'
 
-import CustomPinScreen from './CustomPinScreen'
+import {CustomPinScreen} from './CustomPinScreen'
 
 storiesOf('CustomPinScreen', module).add('Default', () => {
   const route = {
