+++ src/Settings/CustomPin/index.ts
@@ -1,0 +1,1 @@
+export * from './CustomPinScreen'
