+++ src/Settings/SettingsScreenNavigator.tsx
@@ -3,7 +3,6 @@ import {createStackNavigator} from '@react-navigation/stack'
 import React from 'react'
 import {defineMessages, useIntl} from 'react-intl'
 
-import CustomPinScreen from '../../legacy/components/FirstRun/CustomPinScreen'
 import BiometricAuthScreen from '../../legacy/components/Send/BiometricAuthScreen'
 import BiometricsLinkScreen from '../../legacy/components/Settings/BiometricsLinkScreen'
 import ChangeCustomPinScreen from '../../legacy/components/Settings/ChangeCustomPinScreen'
@@ -17,6 +16,7 @@ import {COLORS} from '../../legacy/styles/config'
 import {ApplicationSettingsScreen} from './ApplicationSettings'
 import {ChangePasswordScreen} from './ChangePassword/ChangePasswordScreen'
 import {ChangeWalletName} from './ChangeWalletName/ChangeWalletName'
+import {CustomPinScreen} from './CustomPin'
 import {RemoveWalletScreen} from './RemoveWallet'
 import {WalletSettingsScreen} from './WalletSettings'
